+++ biocpd/__init__.py
@@ -15,4 +15,22 @@ from .affine_registration import AffineRegistration
 from .deformable_registration import DeformableRegistration
 from .constrained_deformable_registration import ConstrainedDeformableRegistration
 from .atlas_registration import AtlasRegistration
+from .pose_marginalized import (
+    PoseMarginalizedConfig,
+    PoseMarginalizedInitialization,
+    pose_marginalized_initialization,
+)
 from .utility import *
+
+__all__ = [
+    "AffineRegistration",
+    "AtlasRegistration",
+    "ConstrainedDeformableRegistration",
+    "DeformableRegistration",
+    "PoseMarginalizedConfig",
+    "PoseMarginalizedInitialization",
+    "RigidRegistration",
+    "gaussian_kernel",
+    "is_positive_semi_definite",
+    "pose_marginalized_initialization",
+]